Adding option to _write and write methods.
The option is called `slim` and default value is False.
When slim is False all ROS data is written to file.
When slim is True all ROS data but the greedy errors
are written to file. Usually, slim should be False since
the error arrays do not take so much space but if memory
space and file size are important that set slim to True.