Removed seeds options from functions.
It does not seem to make sense to rigidly specify the
initial set of seed points since the data is partitioned
and then randomly distributed into these partitions so
that a seed point initially given might be selected
for a validation partition and then can't be used as
as seed.

If a seed point does get selected for a validation partition
then I could choose the nearest point in the full data set
that is not in the validation set but this gets a bit tricky
and, in any case, would not correspond to the actual
seed choice being requested. So, it's probably best
not to allow for the user to specify an initial seed.