Adding example IPython notebook for reduced-order spline error estimation.
Notebook shows how to use RomSpline to:
1) determine if a reduced-order spline is convergent for the given
tolerance and data set;
2) see how robust the size of the reduced data set is to the initial
choice of seed points used in the greedy algorithm that built the
reduced-order spline;
3) build a 'smallest' spline, which is the one with the smallest
reduced data set found after sampling in polynomial degrees and
initial seeds for the greedy algorithm;
4) perform cross-validation studies to estimate an upper bound
for the global reduced-order spline interpolation error for
predicting new values not originally in the interpolated data set.

The cross-validation studies shown include leave-one-out (using the
LeaveOneOut method of the CrossValidation class), K-fold cross-
validation (using the Kfold method), and Monte Carlo K-fold cross-
validation (using the MonteCarloKfold method). See the documentation
for further details about what these studies are and how to use
these cross-validation methods.

Finally, an example is given for estimating the interpolation error
when the requested tolerance is too high so that the spline is not
in a convergent regime.